refactor: 상품 저장소를 새로운 객체로 분리
